% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojo.R
\name{create_mojo}
\alias{create_mojo}
\title{Make Mojo boxoffice data}
\usage{
create_mojo()
}
\description{
make mojo boxoffice data. Example url: http://www.boxofficemojo.com/weekend/chart/?view=&yr=2016&wknd=02&p=.htm
}

