% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball.R
\name{get_ncaa}
\alias{get_ncaa}
\title{Get all NCAA data}
\usage{
get_ncaa(start = 2002, end = 2017)
}
\arguments{
\item{start}{year to start with}

\item{end}{year to end with}
}
\description{
Starting at 2002, and ends with 2017
}
